identified tests which must run in sequence
