+++ Tests.Integration.Presentation.Web/Deltas/V2/DeltaFeedApiV2Test.cs
@@ -14,11 +14,13 @@ using Presentation.Web.Models.API.V2.Response.SystemUsage;
 using Presentation.Web.Models.API.V2.Types.Shared;
 using Tests.Integration.Presentation.Web.Tools;
 using Tests.Integration.Presentation.Web.Tools.External;
+using Tests.Integration.Presentation.Web.Tools.XUnit;
 using Tests.Toolkit.Patterns;
 using Xunit;
 
 namespace Tests.Integration.Presentation.Web.Deltas.V2
 {
+    [Collection(nameof(SequentialTestGroup))]
     public class DeltaFeedApiV2Test : WithAutoFixture
     {
         [Fact]
